Sukurtas atskiras .cpp failas funkcijoms saugoti
